% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPresentation.R
\name{as.character.sigr_statistic}
\alias{as.character.sigr_statistic}
\title{as.character}
\usage{
\method{as.character}{sigr_statistic}(x, ...)
}
\arguments{
\item{x}{sigr wrapper to print}

\item{...}{extra arguments for sigr::render}
}
\value{
formatted string
}
\description{
as.character
}
\examples{

as.character(wrapSignificance(1/300))

}
