% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{build_ROC_curve}
\alias{build_ROC_curve}
\title{calculate ROC curve.}
\usage{
build_ROC_curve(modelPredictions, yValues, ..., na.rm = FALSE, yTarget = TRUE)
}
\arguments{
\item{modelPredictions}{numeric predictions (not empty)}

\item{yValues}{truth values (not empty, same length as model predictions)}

\item{...}{force later arguments to bind by name.}

\item{na.rm}{logical, if TRUE remove NA values.}

\item{yTarget}{value considered to be positive.}
}
\value{
the ROC graph of Score (model score), Sensitivity, and Specificity. Guaranteed to have the (0, 0) and (1, 1) (1-Specificity,Sensitivity) endpoints.
}
\description{
Based on:
 \url{https://blog.revolutionanalytics.com/2016/08/roc-curves-in-two-lines-of-code.html}
}
\examples{

sigr::build_ROC_curve(1:4, c(TRUE,FALSE,TRUE,TRUE))

}
