% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{calcDeviance}
\alias{calcDeviance}
\title{Calculate deviance.}
\usage{
calcDeviance(pred, y)
}
\arguments{
\item{pred}{numeric predictions}

\item{y}{logical truth}
}
\value{
deviance
}
\description{
Calculate deviance.
}
\examples{

sigr::calcDeviance(1:4,c(TRUE,FALSE,TRUE,TRUE))

}
