% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{calcSSE}
\alias{calcSSE}
\title{Calculate sum of squared error.}
\usage{
calcSSE(pred, y, na.rm = FALSE)
}
\arguments{
\item{pred}{numeric predictions}

\item{y}{numeric truth}

\item{na.rm}{logical, if TRUE remove NA values}
}
\value{
sum of squared error
}
\description{
Calculate sum of squared error.
}
\examples{

sigr::calcSSE(1:4,c(1,0,1,1))

}
