% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{fit_beta_shapes}
\alias{fit_beta_shapes}
\title{Fit beta parameters from data.}
\usage{
fit_beta_shapes(x)
}
\arguments{
\item{x}{numeric predictions}
}
\value{
beta shape1, shape2 parameters in a named list
}
\description{
Fit shape1, shape2 using the method of moments.
}
\examples{

x <- rbeta(1000, shape1 = 3, shape2 = 5.5)
fit_beta_shapes(x) # should often be near [3, 5.5]

}
