% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPresentation.R
\name{format.sigr_statistic}
\alias{format.sigr_statistic}
\title{Format}
\usage{
\method{format}{sigr_statistic}(x, ...)
}
\arguments{
\item{x}{sigr wrapper to print}

\item{...}{extra arguments for sigr::render}
}
\value{
formatted string
}
\description{
Format
}
\examples{

format(wrapSignificance(1/300))

}
