% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCTest.R
\name{formatAUCresample}
\alias{formatAUCresample}
\title{Format AUC resampling test results.}
\usage{
formatAUCresample(d, modelName, yName, yTarget, ..., format,
  returnScores = FALSE, nrep = 100, pLargeCutoff = 0.05,
  pSmallCutoff = 1e-05, parallelCluster = NULL)
}
\arguments{
\item{d}{data.frame}

\item{modelName}{character model column name}

\item{yName}{character outcome column name}

\item{yTarget}{target to match to y}

\item{...}{extra arguments (not used)}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii", "docx", ...)}

\item{returnScores}{logical if TRUE return detailed resampledScores.}

\item{nrep}{number of permutation repititions to estimate p values.}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}

\item{parallelCluster}{(optional) a cluster object created by package parallel or package snow.}
}
\value{
formatted string and fields
}
\description{
Estimate signficance of AUC by permuation test.
}
\examples{

set.seed(25325)
d <- data.frame(x1=c(1,2,3,4,5,6,7,7),
                y=c(FALSE,TRUE,FALSE,FALSE,
                    TRUE,TRUE,FALSE,TRUE))
sigr::formatAUCresample(d,'x1','y',TRUE)$formatStr


}

