% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{formatChiSqTest.data.frame}
\alias{formatChiSqTest.data.frame}
\title{Format ChiSqTest from data.}
\usage{
\method{formatChiSqTest}{data.frame}(x, predictionColumnName, yColumnName,
  nParameters = 1, meany = mean(x[[yColumnName]]), ..., format,
  pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{x}{data frame containing columns to compare}

\item{predictionColumnName}{character name of prediction column}

\item{yColumnName}{character name of column containing dependent variable}

\item{nParameters}{number of variables in model}

\item{meany}{(optional) mean of y}

\item{...}{extra arguments (not used)}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format ChiSqTest from data.
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
      y=c(TRUE,FALSE,FALSE,FALSE,TRUE,TRUE,TRUE,FALSE))
model <- glm(y~x,data=d,family=binomial)
summary(model)
d$pred <- predict(model,type='response',newdata=d)
sigr::formatChiSqTest(d,'pred','y',pLargeCutoff=1)$formatStr

}

