% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{formatChiSqTest.glm}
\alias{formatChiSqTest.glm}
\title{Format ChiSqTest from model.}
\usage{
\method{formatChiSqTest}{glm}(x, ..., format, pLargeCutoff = 0.05,
  pSmallCutoff = 1e-05)
}
\arguments{
\item{x}{glm logistic regression model}

\item{...}{extra arguments (not used)}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format ChiSqTest from model.
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
      y=c(TRUE,FALSE,FALSE,FALSE,TRUE,TRUE,TRUE,FALSE))
model <- glm(y~x,data=d,family=binomial)
summary(model)
sigr::formatChiSqTest(model,pLargeCutoff=1,format='ascii')$formatStr


}

