% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{formatChiSqTest}
\alias{formatChiSqTest}
\title{Format quality of a linear regression roughly in "APA Style" ( American Psychological Association ).}
\usage{
formatChiSqTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or lm where to get model or data to score.}

\item{...}{extra arguments}
}
\description{
Format quality of a linear regression roughly in "APA Style" ( American Psychological Association ).
}
\seealso{
\code{\link{formatChiSqTestImpl}},  \code{\link{formatChiSqTest.glm}},  and \code{\link{formatChiSqTest.data.frame}}
}

