% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{formatChiSqTestImpl}
\alias{formatChiSqTestImpl}
\title{Format quality of a logistic regression roughly in "APA Style"
 ( American Psychological Association ).}
\usage{
formatChiSqTestImpl(df.null, df.residual, null.deviance, deviance, format,
  pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{df.null}{null degrees of freedom.}

\item{df.residual}{residual degrees of freedom.}

\item{null.deviance}{null deviance}

\item{deviance}{residual deviance}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii", "docx", ...)}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format quality of a logistic regression roughly in "APA Style"
 ( American Psychological Association ).
}
\examples{

cat(sigr::formatChiSqTestImpl(df.null=7,df.residual=6,
    null.deviance=11.09035,deviance=10.83726)$formatStr)

}

