% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{formatFisherTest}
\alias{formatFisherTest}
\title{Format fisher.test.}
\usage{
formatFisherTest(ft, ..., format, pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{ft}{cor.test result}

\item{...}{extra arguments (not used)}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii", "docx", ...)}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string and fields
}
\description{
Format fisher.test.
}
\examples{

d <- data.frame(x=c('b','a','a','a','b','b','b'),
                y=c('1','1','1','2','2','2','2'))
ft <- fisher.test(table(d))
sigr::formatFisherTest(ft)$formatStr


}

