% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{formatFTest}
\alias{formatFTest}
\title{Format quality of a linear regression roughly in "APA Style" ( American Psychological Association ).}
\usage{
formatFTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or lm where to get model or data to score.}

\item{...}{extra arguments}
}
\description{
Format quality of a linear regression roughly in "APA Style" ( American Psychological Association ).
}
\seealso{
\code{\link{formatFTestImpl}},  \code{\link{formatFTest.lm}},  and \code{\link{formatFTest.data.frame}}
}

