% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{formatFTestImpl}
\alias{formatFTestImpl}
\title{Format quality of a linear regression roughly in "APA Style" ( American Psychological Association ).}
\usage{
formatFTestImpl(numdf, dendf, FValue, format, pLargeCutoff = 0.05,
  pSmallCutoff = 1e-05)
}
\arguments{
\item{numdf}{degrees of freedom 1.}

\item{dendf}{degrees of freedom 2.}

\item{FValue}{observe F test statistic}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format quality of a linear regression roughly in "APA Style" ( American Psychological Association ).
}
\examples{

cat(sigr::formatFTestImpl(numdf=2,dendf=55,FValue=5.56)$formatStr)

}

