% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{formatSignificance}
\alias{formatSignificance}
\title{Print a significance neatly}
\usage{
formatSignificance(sig, symbol = "p", format, pLargeCutoff = 0.05,
  pSmallCutoff = 1e-05)
}
\arguments{
\item{sig}{numeric the significance value.}

\item{symbol}{the name of the value (e.g. p, t, ...).}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Print a significance neatly
}
\examples{

cat(sigr::formatSignificance(1/300))

}

