% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest.R
\name{formatTTest}
\alias{formatTTest}
\title{Format t.test.}
\usage{
formatTTest(tt, ..., format, pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{tt}{t.test result}

\item{...}{extra arguments (not used)}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii", "docx", ...)}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string and fields
}
\description{
Format t.test.
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
tt <- t.test(d$x,d$y)
sigr::formatTTest(tt)$formatStr


}

