% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{getRenderingFormat}
\alias{getRenderingFormat}
\title{Detect rendering format (using knitr).}
\usage{
getRenderingFormat()
}
\value{
rendering format
}
\description{
Detect rendering format (using knitr).
}
\examples{

getRenderingFormat()

}

