% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPresentation.R
\name{print.sigr_statistic}
\alias{print.sigr_statistic}
\title{Print}
\usage{
\method{print}{sigr_statistic}(x, ...)
}
\arguments{
\item{x}{sigr wrapper to print}

\item{...}{extra arguments for sigr::render and print}
}
\value{
formatted string
}
\description{
Print
}
\examples{

print(wrapSignificance(1/300))

}
