% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPresentation.R
\name{render}
\alias{render}
\title{Format summary roughly in "APA Style" ( American Psychological Association ).}
\usage{
render(x, ...)
}
\arguments{
\item{...}{extra arguments}

\item{statistic}{sigr summary statistic}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{sigDigits}{integer number of digits to show}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format summary roughly in "APA Style" ( American Psychological Association ).
}
\seealso{
\code{\link{render.sigr_significance}}, \code{\link{render.sigr_ftest}}
}

