% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCTest.R
\name{render.sigr_aucpairtest}
\alias{render.sigr_aucpairtest}
\title{Format an AUC-test (quality of a probability score)}
\usage{
\method{render}{sigr_aucpairtest}(statistic, ..., format, statDigits = 4,
  sigDigits = 4, pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{statistic}{wrapped AUC test}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{statDigits}{integer number of digits to show in summaries.}

\item{sigDigits}{integer number of digits to show in significances.}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format an AUC-test (quality of a probability score)
}
