% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{render.sigr_chisqtest}
\alias{render.sigr_chisqtest}
\title{Format a chi-square test (quality of categorical prediction)}
\usage{
\method{render}{sigr_chisqtest}(statistic, ..., format, sigDigits = 2,
  pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{statistic}{wrapped T-test}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{sigDigits}{integer number of digits to show}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format a chi-square test (quality of categorical prediction)
}
