% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohenD.R
\name{render.sigr_cohend}
\alias{render.sigr_cohend}
\title{Format Cohen-D (effect size between groups)}
\usage{
\method{render}{sigr_cohend}(statistic, ..., format, statDigits = 3,
  sigDigits = 2, pLargeCutoff = 1, pSmallCutoff = 0)
}
\arguments{
\item{statistic}{wrapped effsize::cohen.d}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{statDigits}{integer number of digits to show in summaries (not yet implemented).}

\item{sigDigits}{integer number of digits to show in significances.}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format Cohen-D (effect size between groups)
}
