% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{render.sigr_ftest}
\alias{render.sigr_ftest}
\title{Format an F-test}
\usage{
\method{render}{sigr_ftest}(statistic, ..., format, statDigits = 3,
  sigDigits = 2, pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{statistic}{wrapped test}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{statDigits}{integer number of digits to show in summaries.}

\item{sigDigits}{integer number of digits to show in significances.}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format an F-test
}
