% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{render.sigr_pwr_htest}
\alias{render.sigr_pwr_htest}
\title{Format a pwr-test}
\usage{
\method{render}{sigr_pwr_htest}(
  statistic,
  ...,
  format,
  statDigits = 4,
  sigDigits = 4,
  pLargeCutoff = 1,
  pSmallCutoff = 1e-05
)
}
\arguments{
\item{statistic}{wrapped test from pwr package}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{statDigits}{integer number of digits to show in summaries.}

\item{sigDigits}{integer number of digits to show in significances.}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format a pwr-test
}
