% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPresentation.R
\name{render.sigr_significance}
\alias{render.sigr_significance}
\title{Format a significance}
\usage{
\method{render}{sigr_significance}(statistic, ..., format, sigDigits = 2,
  pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{statistic}{wrapped significance}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{sigDigits}{integer number of digits to show}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format a significance
}
\examples{

cat(render(wrapSignificance(1/300),format='html'))

}

