% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest.R
\name{render.sigr_ttest}
\alias{render.sigr_ttest}
\title{Format a T-test (difference in means by group)}
\usage{
\method{render}{sigr_ttest}(statistic, ..., format, statDigits = 2,
  sigDigits = 2, pLargeCutoff = 0.05, pSmallCutoff = 1e-05)
}
\arguments{
\item{statistic}{wrapped T-test}

\item{...}{not used, force use of named binding for later arguments}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii")}

\item{statDigits}{integer number of digits to show in summaries (not yet implemented).}

\item{sigDigits}{integer number of digits to show in significances.}

\item{pLargeCutoff}{value to declare non-significance at or above.}

\item{pSmallCutoff}{smallest value to print}
}
\value{
formatted string
}
\description{
Format a T-test (difference in means by group)
}
