% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{sensitivity_from_specificity_q}
\alias{sensitivity_from_specificity_q}
\title{Compute the q-graph.}
\usage{
sensitivity_from_specificity_q(Specificity, q)
}
\arguments{
\item{Specificity}{vector of sensitivities to evaluate}

\item{q}{shape parameter for \code{1 - (1 - (1-Specificity)^q)^(1/q)}}
}
\value{
Sensitivity
}
\description{
Based on:
 \url{https://blog.revolutionanalytics.com/2016/08/roc-curves-in-two-lines-of-code.html}
}
\examples{

sensitivity_from_specificity_q(seq(0, 1, 0.1), 0.61)

}
