% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigr.R
\docType{package}
\name{sigr}
\alias{sigr}
\alias{sigr-package}
\title{sigr: Format Significance Summaries for Reports}
\description{
Succinctly format significance summaries of
various models and tests.  The main purpose is unified reporting and planning
of experimental results, working around issue such as the difficulty of
extracting model summary facts (such as with 'lm'/'glm').  This package also
includes empirical tests, such as bootstrap estimates.
}
\details{
To learn more about replyr, please start with the vignette:
\code{vignette('sigrFormatting','sigr')}
}

