% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TInterval.R
\name{TInterval}
\alias{TInterval}
\title{Wrap TInterval (test of Binomial/Bernoulli rate).}
\usage{
TInterval(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or test.}

\item{...}{extra arguments}
}
\description{
Wrap TInterval (test of Binomial/Bernoulli rate).
}
\seealso{
\code{\link{TIntervalS}}, \code{\link{TInterval.numeric}}, \code{\link{TInterval.data.frame}}
}
