% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{wrapChiSqTest.glm}
\alias{wrapChiSqTest.glm}
\title{Format ChiSqTest from model.}
\usage{
\method{wrapChiSqTest}{glm}(x, ...)
}
\arguments{
\item{x}{glm logistic regression model}

\item{...}{extra arguments (not used)}
}
\value{
wrapped test
}
\description{
Format ChiSqTest from model.
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
      y=c(TRUE,FALSE,FALSE,FALSE,TRUE,TRUE,TRUE,FALSE))
model <- glm(y~x,data=d,family=binomial)
summary(model)
render(wrapChiSqTest(model),pLargeCutoff=1,format='ascii')


}

