% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{wrapChiSqTest}
\alias{wrapChiSqTest}
\title{Wrap quality of a categorical prediction roughly in "APA Style" ( American Psychological Association ).}
\usage{
wrapChiSqTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or lm where to get model or data to score.}

\item{...}{extra arguments}
}
\description{
Wrap quality of a categorical prediction roughly in "APA Style" ( American Psychological Association ).
}
\seealso{
\code{\link{wrapChiSqTestImpl}},  \code{\link{wrapChiSqTest.glm}},  and \code{\link{wrapChiSqTest.data.frame}}
}

