% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{wrapChiSqTest.summary.glm}
\alias{wrapChiSqTest.summary.glm}
\title{Format ChiSqTest from model summary.}
\usage{
\method{wrapChiSqTest}{summary.glm}(x, ...)
}
\arguments{
\item{x}{summary(glm(family=binomial)) object.}

\item{...}{extra arguments (not used)}
}
\value{
wrapped test
}
\description{
Format ChiSqTest from model summary.
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
      y=c(TRUE,FALSE,FALSE,FALSE,TRUE,TRUE,TRUE,FALSE))
model <- glm(y~x,data=d,family=binomial)
sum <- summary(model)
render(wrapChiSqTest(sum),pLargeCutoff=1,format='ascii')


}
