% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSqTest.R
\name{wrapChiSqTestImpl}
\alias{wrapChiSqTestImpl}
\title{Format quality of a logistic regression roughly in "APA Style"
 ( American Psychological Association ).}
\usage{
wrapChiSqTestImpl(df.null, df.residual, null.deviance, deviance)
}
\arguments{
\item{df.null}{null degrees of freedom.}

\item{df.residual}{residual degrees of freedom.}

\item{null.deviance}{null deviance}

\item{deviance}{residual deviance}
}
\value{
wrapped statistic
}
\description{
Format quality of a logistic regression roughly in "APA Style"
 ( American Psychological Association ).
}
\examples{

wrapChiSqTestImpl(df.null=7,df.residual=6,
    null.deviance=11.09035,deviance=10.83726)

}
