% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohenD.R
\name{wrapCohenD.data.frame}
\alias{wrapCohenD.data.frame}
\title{Wrap Cohen's D (effect size between groups).}
\usage{
\method{wrapCohenD}{data.frame}(x, Column1Name, Column2Name, ...,
  pooled = TRUE, paired = FALSE, na.rm = FALSE,
  hedges.correction = FALSE, conf.level = 0.95, noncentral = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{Column1Name}{character column 1 name}

\item{Column2Name}{character column 2 name}

\item{...}{extra arguments passed to effsize::cohen.d}

\item{pooled}{passed to \code{\link[effsize]{cohen.d}}}

\item{paired}{passed to \code{\link[effsize]{cohen.d}}}

\item{na.rm}{passed to \code{\link[effsize]{cohen.d}}}

\item{hedges.correction}{passed to \code{\link[effsize]{cohen.d}}}

\item{conf.level}{passed to \code{\link[effsize]{cohen.d}}}

\item{noncentral}{passed to \code{\link[effsize]{cohen.d}}}
}
\value{
formatted string and fields
}
\description{
Wrap Cohen's D (effect size between groups).
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
render(wrapCohenD(d,'x','y'))

}
