% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohenD.R
\name{wrapCohenD.effsize}
\alias{wrapCohenD.effsize}
\title{Wrap Cohen's D (effect size between groups).}
\usage{
\method{wrapCohenD}{effsize}(x, ...)
}
\arguments{
\item{x}{effsize::cohen.d result}

\item{...}{extra arguments (not used)}
}
\value{
formatted string and fields
}
\description{
Wrap Cohen's D (effect size between groups).
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
cohen_d <- effsize::cohen.d(d$x,d$y)
render(wrapCohenD(cohen_d))

}
