% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohenD.R
\name{wrapCohenD}
\alias{wrapCohenD}
\title{Wrap Cohen's D (effect size between groups).}
\usage{
wrapCohenD(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or test.}

\item{...}{extra arguments}
}
\description{
Wrap Cohen's D (effect size between groups).
}
\seealso{
\code{\link[effsize]{cohen.d}}, \code{\link{wrapCohenD.effsize}}, and \code{\link{wrapCohenD.data.frame}}
}
