% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTest.R
\name{wrapCorTest.data.frame}
\alias{wrapCorTest.data.frame}
\title{Wrap cor.test (test of liner correlation).}
\usage{
\method{wrapCorTest}{data.frame}(x, Column1Name, Column2Name, ...,
  alternative = c("two.sided", "less", "greater"), method = c("pearson",
  "kendall", "spearman"), exact = NULL, conf.level = 0.95,
  continuity = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{Column1Name}{character column 1 name}

\item{Column2Name}{character column 2 name}

\item{...}{extra arguments passed to cor.test}

\item{alternative}{passed to \code{\link[stats]{cor.test}}}

\item{method}{passed to \code{\link[stats]{cor.test}}}

\item{exact}{passed to \code{\link[stats]{cor.test}}}

\item{conf.level}{passed to \code{\link[stats]{cor.test}}}

\item{continuity}{passed to \code{\link[stats]{cor.test}}}

\item{na.rm}{logical, if TRUE remove NA values}
}
\value{
wrapped stat
}
\description{
Wrap cor.test (test of liner correlation).
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
wrapCorTest(d,'x','y')


}
