% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTest.R
\name{wrapCorTest.htest}
\alias{wrapCorTest.htest}
\title{Wrap cor.test (test of liner correlation).}
\usage{
\method{wrapCorTest}{htest}(x, ...)
}
\arguments{
\item{x}{cor.test result}

\item{...}{extra arguments (not used)}
}
\value{
wrapped stat
}
\description{
Wrap cor.test (test of liner correlation).
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
ct <- cor.test(d$x,d$y)
wrapCorTest(ct)


}
