% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTest.R
\name{wrapCorTest}
\alias{wrapCorTest}
\title{Wrap cor.test (test of liner correlation).}
\usage{
wrapCorTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or test.}

\item{...}{extra arguments}
}
\description{
Wrap cor.test (test of liner correlation).
}
\seealso{
\code{\link{wrapCorTest.htest}}, and  \code{\link{wrapCorTest.data.frame}}
}

