% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{wrapFisherTest.data.frame}
\alias{wrapFisherTest.data.frame}
\title{Wrap fisher.test (test of categorial indendence).}
\usage{
\method{wrapFisherTest}{data.frame}(x, Column1Name, Column2Name, ...)
}
\arguments{
\item{x}{data.frame}

\item{Column1Name}{character column 1 name}

\item{Column2Name}{character column 2 name}

\item{...}{extra arguments}
}
\value{
wrapped test.
}
\description{
Wrap fisher.test (test of categorial indendence).
}
\examples{

d <- data.frame(x=c('b','a','a','a','b','b','b'),
                y=c('1','1','1','2','2','2','2'))
wrapFisherTest(d,'x','y')


}

