% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{wrapFisherTest.htest}
\alias{wrapFisherTest.htest}
\title{Wrap fisher.test (test of categorial indendence).}
\usage{
\method{wrapFisherTest}{htest}(x, ...)
}
\arguments{
\item{x}{fisher.test result}

\item{...}{extra arguments (not used)}
}
\value{
wrapped test.
}
\description{
Wrap fisher.test (test of categorial indendence).
}
\examples{

d <- data.frame(x=c('b','a','a','a','b','b','b'),
                y=c('1','1','1','2','2','2','2'))
ft <- fisher.test(table(d))
wrapFisherTest(ft)


}

