% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{wrapFisherTest}
\alias{wrapFisherTest}
\title{Wrap fisher.test (test of categorical independence).}
\usage{
wrapFisherTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or test.}

\item{...}{extra arguments}
}
\description{
Wrap fisher.test (test of categorical independence).
}
\seealso{
\code{\link{wrapFisherTest.htest}}, and  \code{\link{wrapFisherTest.data.frame}}
}
