% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{wrapFTest.htest}
\alias{wrapFTest.htest}
\title{Wrap F-test (ratio of variances).}
\usage{
\method{wrapFTest}{htest}(x, ..., format = NULL)
}
\arguments{
\item{x}{lm model}

\item{...}{extra arguments (not used)}

\item{format}{if set the format to return ("html", "latex", "markdown", "ascii", "docx", ...)}
}
\value{
formatted string
}
\description{
Wrap F-test (ratio of variances).
}
\examples{

v <- var.test(c(1,2,3,4,5,6,7,7), c(1, 1, 2))
sigr::wrapFTest(v)


}
