% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{wrapFTest}
\alias{wrapFTest}
\title{Wrap F-test (significance identity relation).}
\usage{
wrapFTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or lm where to get model or data to score.}

\item{...}{extra arguments}
}
\description{
Wrap F-test (significance identity relation).
}
\seealso{
\code{\link{wrapFTestImpl}},  \code{\link{wrapFTest.lm}},  and \code{\link{wrapFTest.data.frame}}
}
