% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{wrapFTestezANOVA}
\alias{wrapFTestezANOVA}
\title{Wrap quality statistic of a linear relation from ezANOVA (package ez).}
\usage{
wrapFTestezANOVA(x, ...)
}
\arguments{
\item{x}{list result from ezANOVA (package ez).}

\item{...}{extra arguments (not used)}
}
\value{
list of formatted string and fields
}
\description{
Please see \url{https://github.com/WinVector/sigr/issues/1#issuecomment-322311947} for an example.
}
