% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{wrapPWR.power.htest}
\alias{wrapPWR.power.htest}
\title{Wrap pwr test.}
\usage{
\method{wrapPWR}{power.htest}(x, ...)
}
\arguments{
\item{x}{pwr test result}

\item{...}{extra arguments (not used)}
}
\value{
formatted string and fields
}
\description{
Wrap pwr test.
}
\examples{

# Example from pwr package
#  Exercise 6.1 p. 198 from Cohen (1988)
test <- pwr::pwr.2p.test(h=0.3,n=80,sig.level=0.05,alternative="greater")
wrapPWR(test)

}
