% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{wrapPWR}
\alias{wrapPWR}
\title{Wrap pwr test (difference in means by group).}
\usage{
wrapPWR(x, ...)
}
\arguments{
\item{x}{test from pwr package}

\item{...}{extra arguments}
}
\description{
Wrap pwr test (difference in means by group).
}
\seealso{
\code{\link[pwr]{pwr.2p.test}}
}
