% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPresentation.R
\name{wrapSignificance}
\alias{wrapSignificance}
\title{Wrap a significance}
\usage{
wrapSignificance(significance, symbol = "p")
}
\arguments{
\item{significance}{numeric the significance value.}

\item{symbol}{the name of the value (e.g. "p", "t", ...).}
}
\value{
wrapped significance
}
\description{
Wrap a significance
}
\examples{

wrapSignificance(1/300)

}
