% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest.R
\name{wrapTTest.data.frame}
\alias{wrapTTest.data.frame}
\title{Wrap t.test (difference in means by group).}
\usage{
\method{wrapTTest}{data.frame}(x, Column1Name, Column2Name, ...,
  na.rm = FALSE)
}
\arguments{
\item{x}{data.frame}

\item{Column1Name}{character column 1 name}

\item{Column2Name}{character column 2 name}

\item{...}{extra arguments passed to ttest}

\item{na.rm}{logical, if TRUE remove NA values}
}
\value{
formatted string and fields
}
\description{
Wrap t.test (difference in means by group).
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
render(wrapTTest(d,'x','y'),pLargeCutoff=1)
# confirm p not order depedent
render(wrapTTest(d,'y','x'),pLargeCutoff=1)

}
