% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest.R
\name{wrapTTest}
\alias{wrapTTest}
\title{Wrap t.test (difference in means by group).}
\usage{
wrapTTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or test.}

\item{...}{extra arguments}
}
\description{
Wrap t.test (difference in means by group).
}
\seealso{
\code{\link{wrapTTest.htest}}, and  \code{\link{wrapTTest.data.frame}}
}

