% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferenceDistribution.R
\name{Bernoulli_diff_dist}
\alias{Bernoulli_diff_dist}
\title{Compute the distribution of differences of replacement samples of two Bernoulli experiments.}
\usage{
Bernoulli_diff_dist(nA, kA, nB, kB)
}
\arguments{
\item{nA}{number of A experiments.}

\item{kA}{number of A successes observed.}

\item{nB}{number of B experiments (must be no larger than nA).}

\item{kB}{number of B successes observed.}
}
\value{
experiment summaries
}
\description{
Compute the distribution of differences of replacement samples of two Bernoulli experiments.
}
\examples{

Bernoulli_diff_dist(5, 2, 2, 1)

}
