% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{calcAUC}
\alias{calcAUC}
\title{calculate AUC.}
\usage{
calcAUC(modelPredictions, yValues, ..., na.rm = FALSE, yTarget = TRUE)
}
\arguments{
\item{modelPredictions}{numeric predictions (not empty)}

\item{yValues}{truth values (not empty, same length as model predictions)}

\item{...}{force later arguments to bind by name.}

\item{na.rm}{logical, if TRUE remove NA values.}

\item{yTarget}{value considered to be positive.}
}
\value{
area under curve
}
\description{
Based on:
 \url{http://blog.revolutionanalytics.com/2016/08/roc-curves-in-two-lines-of-code.html}
}
\examples{

sigr::calcAUC(1:4,c(TRUE,FALSE,TRUE,TRUE)) # should be 2/3

}
