% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{calcDeviance}
\alias{calcDeviance}
\title{Calculate deviance.}
\usage{
calcDeviance(pred, y, na.rm = FALSE, eps = 1e-06)
}
\arguments{
\item{pred}{numeric predictions}

\item{y}{logical truth}

\item{na.rm}{logical, if TRUE remove NA values}

\item{eps}{numeric, smoothing term}
}
\value{
deviance
}
\description{
Calculate deviance.
}
\examples{

sigr::calcDeviance(1:4,c(TRUE,FALSE,TRUE,TRUE))

}
