% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{calcSSE}
\alias{calcSSE}
\title{Calculate sum of squared error.}
\usage{
calcSSE(pred, y)
}
\arguments{
\item{pred}{numeric predictions}

\item{y}{numeric truth}
}
\value{
sum of squared error
}
\description{
Calculate sum of squared error.
}
\examples{

sigr::calcSSE(1:4,c(TRUE,FALSE,TRUE,TRUE))

}

