% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{find_area_q}
\alias{find_area_q}
\title{Find area matching polynomial curve.}
\usage{
find_area_q(area, ..., n_points = 101)
}
\arguments{
\item{area}{area to match}

\item{...}{not used, force later arguments to bind by name}

\item{n_points}{how many points to use to estimte area.}
}
\value{
q that such that curve \code{1 - (1 - (1-Specificity)^q)^(1/q)} matches area
}
\description{
Based on \url{https://win-vector.com/2020/09/13/why-working-with-auc-is-more-powerful-than-one-might-think/}
}
\examples{

find_area_q(0.75)

}
