% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigr.R
\docType{package}
\name{sigr-package}
\alias{sigr}
\alias{sigr-package}
\title{sigr: Format Significance Summaries for Reports}
\description{
Succinctly format significance summaries of
various models and tests (F-test, Chi-Sq-test, Fisher-test, T-test, and rank-significance).  The main purpose is unified reporting and planning
of experimental results, working around issue such as the difficulty of
extracting model summary facts (such as with 'lm'/'glm').  This package also
includes empirical tests, such as bootstrap estimates.
}
\details{
To learn more about sigr, please start with the vignette:
\code{vignette('sigrFormatting','sigr')}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WinVector/sigr/}
  \item \url{https://winvector.github.io/sigr/}
  \item Report bugs at \url{https://github.com/WinVector/sigr/issues}
}

}
\author{
\strong{Maintainer}: John Mount \email{jmount@win-vector.com}

Authors:
\itemize{
  \item Nina Zumel \email{nzumel@win-vector.com}
}

Other contributors:
\itemize{
  \item  Win-Vector LLC [copyright holder]
}

}
