% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomTest.R
\name{wrapBinomTest.htest}
\alias{wrapBinomTest.htest}
\title{Wrap binom.test (test of Bernoulli rate).}
\usage{
\method{wrapBinomTest}{htest}(x, ...)
}
\arguments{
\item{x}{binom.test result}

\item{...}{extra arguments (not used)}
}
\value{
wrapped stat
}
\description{
Wrap binom.test (test of Bernoulli rate).
}
\examples{

bt <- binom.test(7, 10, 0.5)
wrapBinomTest(bt)


}
