% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomTest.R
\name{wrapBinomTest}
\alias{wrapBinomTest}
\title{Wrap binom.test (test of Bernoulli rate).}
\usage{
wrapBinomTest(x, ...)
}
\arguments{
\item{x}{numeric, data.frame or test.}

\item{...}{extra arguments}
}
\description{
Wrap binom.test (test of Bernoulli rate).
}
\seealso{
\code{\link{wrapBinomTest.htest}}, and  \code{\link{wrapBinomTest.data.frame}}
}
