% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTest.R
\name{wrapFTestImpl}
\alias{wrapFTestImpl}
\title{Wrap F-test (significane of a linear relation).}
\usage{
wrapFTestImpl(numdf, dendf, FValue)
}
\arguments{
\item{numdf}{degrees of freedom 1.}

\item{dendf}{degrees of freedom 2.}

\item{FValue}{observed F test statistic}
}
\value{
wrapped statistic
}
\description{
Wrap F-test (significane of a linear relation).
}
\examples{

wrapFTestImpl(numdf=2,dendf=55,FValue=5.56)

}
