% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTest.R
\name{wrapTTest}
\alias{wrapTTest}
\title{Wrap t.test (difference in means by group).}
\usage{
wrapTTest(tt, ...)
}
\arguments{
\item{tt}{t.test result}

\item{...}{extra arguments (not used)}
}
\value{
formatted string and fields
}
\description{
Wrap t.test (difference in means by group).
}
\examples{

d <- data.frame(x=c(1,2,3,4,5,6,7,7),
                y=c(1,1,2,2,3,3,4,4))
tt <- t.test(d$x,d$y)
render(wrapTTest(tt),pLargeCutoff=1)
# confirm not rescaling, as a correlation test would
render(wrapTTest(t.test(d$x,2*d$y)),pLargeCutoff=1)

}

